+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-04-16
+
+### Added
+
+*None*
+
+### Changed
+
+* `InMemory` cache by default in `HttpClient`.
+* `HttpClient` accepts `Amp\Http\Client\DelegateHttpClient` interface.
+
+### Deprecated
+
+*None*
+
+### Removed
+
+*None*
+
+### Fixed
+
+* Fix serialization for large string which contains broken UTF-8 symbols ("Malformed UTF-8 characters, possibly incorrectly encoded).
+
+### Security
+
+*None*
+
 ## [1.0.0] - 2020-04-15
 
 ### Added